Further refactored `OptionsMenuBuilder<T>`: primarily moving the event code into `ConfigFileMetadata<T>`
