+++ SMLHelper/Options/Attributes/ConfigFileMetadata.cs
@@ -1,6 +1,8 @@
 ﻿namespace SMLHelper.V2.Options.Attributes
 {
+    using Interfaces;
     using Json;
+    using Oculus.Newtonsoft.Json;
     using QModManager.API;
     using System;
     using System.Collections.Generic;
@@ -9,6 +11,11 @@
     using System.Reflection;
     using UnityEngine;
     using Logger = Logger;
+#if SUBNAUTICA
+    using Text = UnityEngine.UI.Text;
+#elif BELOWZERO
+    using Text = TMPro.TextMeshProUGUI;
+#endif
 
     internal class ConfigFileMetadata<T> where T : ConfigFile, new()
     {
@@ -251,5 +258,372 @@
         {
             return ModOptionAttributesMetadata.Values;
         }
+
+        #region Events
+        public void BindEvents()
+        {
+            Config.OnStartedLoading += OptionsMenuBuilder_Config_OnStartedLoading;
+            Config.OnFinishedLoading += OptionsMenuBuilder_Config_OnFinishedLoading;
+        }
+
+        private string jsonConfig;
+        public void OptionsMenuBuilder_Config_OnStartedLoading(object sender, ConfigFileEventArgs e)
+        {
+            jsonConfig = JsonConvert.SerializeObject(e.Instance as T);
+        }
+
+        private void OptionsMenuBuilder_Config_OnFinishedLoading(object sender, ConfigFileEventArgs e)
+        {
+            T oldConfig = JsonConvert.DeserializeObject<T>(jsonConfig);
+            T currentConfig = e.Instance as T;
+
+            foreach (ModOptionAttributeMetadata<T> modOptionMetadata in GetValues())
+            {
+                if (modOptionMetadata.MemberInfoMetadata.MemberType != MemberType.Field &&
+                    modOptionMetadata.MemberInfoMetadata.MemberType != MemberType.Property)
+                {
+                    continue;
+                }
+
+                if (!modOptionMetadata.MemberInfoMetadata.GetValue(oldConfig)
+                    .Equals(modOptionMetadata.MemberInfoMetadata.GetValue(currentConfig)))
+                {
+                    invokeOnChangeEvents(modOptionMetadata, sender);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Invokes the method for a given <see cref="ButtonAttribute"/> and passes parameters when the button is clicked.
+        /// </summary>
+        /// <param name="sender">The sender of the original button click event.</param>
+        /// <param name="e">The <see cref="ButtonClickedEventArgs"/> for the click event.</param>
+        public void HandleButtonClick(object sender, ButtonClickedEventArgs e)
+        {
+            if (TryGetMetadata(e.Id, out ModOptionAttributeMetadata<T> modOptionMetadata)
+                && modOptionMetadata.MemberInfoMetadata.MethodParameterTypes is Type[] parameterTypes)
+            {
+                var parameters = new object[parameterTypes.Length];
+                var senderFound = false;
+                var eventArgsFound = false;
+
+                for (var i = 0; i < parameters.Length; i++)
+                {
+                    Type type = parameterTypes[i];
+                    if (!senderFound && type == typeof(object))
+                    {
+                        senderFound = true;
+                        parameters[i] = sender;
+                    }
+                    else if (!eventArgsFound && type == typeof(ButtonClickedEventArgs))
+                    {
+                        eventArgsFound = true;
+                        parameters[i] = e;
+                    }
+
+                    if (senderFound && eventArgsFound)
+                        break;
+                }
+
+                modOptionMetadata.MemberInfoMetadata.InvokeMethod(Config, parameters);
+            }
+        }
+
+        /// <summary>
+        /// Sets the value in the <see cref="Config"/>, optionally saving the <see cref="Config"/> to disk if the
+        /// <see cref="MenuAttribute.SaveEvents.ChangeValue"/> flag is set, before passing off to
+        /// <see cref="invokeOnChangeEvents{TSource}(ModOptionAttributeMetadata{T}, object, TSource)"/>
+        /// to invoke any methods specified with an <see cref="OnChangeAttribute"/>.
+        /// </summary>
+        /// <param name="sender">The sender of the original choice changed event.</param>
+        /// <param name="e">The <see cref="ChoiceChangedEventArgs"/> for the choice changed event.</param>
+        public void HandleChoiceChanged(object sender, ChoiceChangedEventArgs e)
+        {
+            if (TryGetMetadata(e.Id, out ModOptionAttributeMetadata<T> modOptionMetadata))
+            {
+                // Set the value in the Config
+                MemberInfoMetadata<T> memberInfoMetadata = modOptionMetadata.MemberInfoMetadata;
+                ChoiceAttribute choiceAttribute = modOptionMetadata.ModOptionAttribute as ChoiceAttribute;
+
+                if (memberInfoMetadata.ValueType.IsEnum && (choiceAttribute.Options == null || !choiceAttribute.Options.Any()))
+                {
+                    // Enum-based choice where the values are parsed from the enum type
+                    object value = Enum.Parse(memberInfoMetadata.ValueType, e.Value);
+                    memberInfoMetadata.SetValue(Config, value);
+                }
+                else if (memberInfoMetadata.ValueType.IsEnum)
+                {
+                    // Enum-based choice where the values are defined as custom strings
+                    object value = Enum.Parse(memberInfoMetadata.ValueType, Enum.GetNames(memberInfoMetadata.ValueType)[e.Index]);
+                    memberInfoMetadata.SetValue(Config, value);
+                }
+                else if (memberInfoMetadata.ValueType == typeof(string))
+                {
+                    // string-based choice value
+                    string value = e.Value;
+                    memberInfoMetadata.SetValue(Config, value);
+                }
+                else if (memberInfoMetadata.ValueType == typeof(int))
+                {
+                    // index-based choice value
+                    int value = e.Index;
+                    memberInfoMetadata.SetValue(Config, value);
+                }
+
+                // Optionally save the Config to disk
+                if (MenuAttribute.SaveOn.HasFlag(MenuAttribute.SaveEvents.ChangeValue))
+                    Config.Save();
+
+                // Invoke any OnChange methods specified
+                invokeOnChangeEvents(modOptionMetadata, sender, e);
+            }
+        }
+
+        /// <summary>
+        /// Sets the value in the <see cref="Config"/>, optionally saving the <see cref="Config"/> to disk if the
+        /// <see cref="MenuAttribute.SaveEvents.ChangeValue"/> flag is set, before passing off to
+        /// <see cref="invokeOnChangeEvents{TSource}(ModOptionAttributeMetadata{T}, object, TSource)"/>
+        /// to invoke any methods specified with an <see cref="OnChangeAttribute"/>.
+        /// </summary>
+        /// <param name="sender">The sender of the original keybind changed event.</param>
+        /// <param name="e">The <see cref="KeybindChangedEventArgs"/> for the keybind changed event.</param>
+        public void HandleKeybindChanged(object sender, KeybindChangedEventArgs e)
+        {
+            if (TryGetMetadata(e.Id, out ModOptionAttributeMetadata<T> modOptionMetadata))
+            {
+                // Set the value in the Config
+                modOptionMetadata.MemberInfoMetadata.SetValue(Config, e.Key);
+
+                // Optionally save the Config to disk
+                if (MenuAttribute.SaveOn.HasFlag(MenuAttribute.SaveEvents.ChangeValue))
+                    Config.Save();
+
+                // Invoke any OnChange methods specified
+                invokeOnChangeEvents(modOptionMetadata, sender, e);
+            }
+        }
+
+        /// <summary>
+        /// Sets the value in the <see cref="Config"/>, optionally saving the <see cref="Config"/> to disk if the
+        /// <see cref="MenuAttribute.SaveEvents.ChangeValue"/> flag is set, before passing off to
+        /// <see cref="invokeOnChangeEvents{TSource}(ModOptionAttributeMetadata{T}, object, TSource)"/>
+        /// to invoke any methods specified with an <see cref="OnChangeAttribute"/>.
+        /// </summary>
+        /// <param name="sender">The sender of the original slider changed event.</param>
+        /// <param name="e">The <see cref="SliderChangedEventArgs"/> for the slider changed event.</param>
+        public void HandleSliderChanged(object sender, SliderChangedEventArgs e)
+        {
+            if (TryGetMetadata(e.Id, out ModOptionAttributeMetadata<T> modOptionMetadata))
+            {
+                // Set the value in the Config
+                MemberInfoMetadata<T> memberInfoMetadata = modOptionMetadata.MemberInfoMetadata;
+                object value = Convert.ChangeType(e.Value, memberInfoMetadata.ValueType);
+                memberInfoMetadata.SetValue(Config, value);
+
+                // Optionally save the Config to disk
+                if (MenuAttribute.SaveOn.HasFlag(MenuAttribute.SaveEvents.ChangeValue))
+                    Config.Save();
+
+                // Invoke any OnChange methods specified
+                invokeOnChangeEvents(modOptionMetadata, sender, e);
+            }
+        }
+
+        /// <summary>
+        /// Sets the value in the <see cref="Config"/>, optionally saving the <see cref="Config"/> to disk if the
+        /// <see cref="MenuAttribute.SaveEvents.ChangeValue"/> flag is set, before passing off to
+        /// <see cref="invokeOnChangeEvents{TSource}(ModOptionAttributeMetadata{T}, object, TSource)"/>
+        /// to invoke any methods specified with an <see cref="OnChangeAttribute"/>.
+        /// </summary>
+        /// <param name="sender">The sender of the original toggle changed event.</param>
+        /// <param name="e">The <see cref="ToggleChangedEventArgs"/> for the toggle changed event.</param>
+        public void HandleToggleChanged(object sender, ToggleChangedEventArgs e)
+        {
+            if (TryGetMetadata(e.Id, out ModOptionAttributeMetadata<T> modOptionMetadata))
+            {
+                // Set the value in the Config
+                modOptionMetadata.MemberInfoMetadata.SetValue(Config, e.Value);
+
+                // Optionally save the Config to disk
+                if (MenuAttribute.SaveOn.HasFlag(MenuAttribute.SaveEvents.ChangeValue))
+                    Config.Save();
+
+                // Invoke any OnChange methods specified
+                invokeOnChangeEvents(modOptionMetadata, sender, e);
+            }
+        }
+
+        /// <summary>
+        /// Generates tooltips for each <see cref="ModOption"/> with a specified <see cref="TooltipAttribute"/>, before
+        /// invoking any relevant method(s) specified with <see cref="OnGameObjectCreatedAttribute"/>(s) and passes
+        /// parameters when a <see cref="GameObject"/> is created in the options menu.
+        /// </summary>
+        /// <param name="sender"></param>
+        /// <param name="e"></param>
+        public void HandleGameObjectCreated(object sender, GameObjectCreatedEventArgs e)
+        {
+            if (TryGetMetadata(e.Id, out ModOptionAttributeMetadata<T> modOptionMetadata))
+            {
+                // Create a tooltip if there is a TooltipAttribute specified
+                if (modOptionMetadata.ModOptionAttribute.Tooltip is string tooltip)
+                {
+                    e.GameObject.GetComponentInChildren<Text>().gameObject.AddComponent<ModOptionTooltip>().Tooltip = tooltip;
+                }
+
+                if (modOptionMetadata.OnGameObjectCreatedMetadata == null)
+                    return; // Skip attempting to invoke events if there are no OnGameObjectCreatedAttributes set for the member.
+
+                foreach (MemberInfoMetadata<T> onGameObjectCreatedMetadata in modOptionMetadata.OnGameObjectCreatedMetadata)
+                {
+                    invokeEvent(onGameObjectCreatedMetadata, sender, e);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Invokes the relevant method(s) specified with the <see cref="OnChangeAttribute"/>(s)
+        /// and passes parameters when a value is changed when loaded from disk.
+        /// </summary>
+        /// <param name="modOptionMetadata">The metadata for the mod option.</param>
+        /// <param name="sender">The sender of the event.</param>
+        private void invokeOnChangeEvents(ModOptionAttributeMetadata<T> modOptionMetadata, object sender)
+        {
+            string id = modOptionMetadata.ModOptionAttribute.Id;
+            MemberInfoMetadata<T> memberInfoMetadata = modOptionMetadata.MemberInfoMetadata;
+
+            switch (modOptionMetadata.ModOptionAttribute)
+            {
+                case ChoiceAttribute choiceAttribute when memberInfoMetadata.ValueType.IsEnum &&
+                    (choiceAttribute.Options == null || !choiceAttribute.Options.Any()):
+                    // Enum-based choice where the values are parsed from the enum type
+                    {
+                        string[] options = Enum.GetNames(memberInfoMetadata.ValueType);
+                        string value = memberInfoMetadata.GetValue(Config).ToString();
+                        var eventArgs = new ChoiceChangedEventArgs(id, Array.IndexOf(options, value), value);
+                        invokeOnChangeEvents(modOptionMetadata, sender, eventArgs);
+                    }
+                    break;
+                case ChoiceAttribute _ when memberInfoMetadata.ValueType.IsEnum:
+                    // Enum-based choice where the values are defined as custom strings
+                    {
+                        string value = memberInfoMetadata.GetValue(Config).ToString();
+                        int index = Math.Max(Array.IndexOf(Enum.GetValues(memberInfoMetadata.ValueType), value), 0);
+                        var eventArgs = new ChoiceChangedEventArgs(id, index, value);
+                        invokeOnChangeEvents(modOptionMetadata, sender, eventArgs);
+                    }
+                    break;
+                case ChoiceAttribute choiceAttribute when memberInfoMetadata.ValueType == typeof(string):
+                    // string-based choice value
+                    {
+                        string[] options = choiceAttribute.Options;
+                        string value = memberInfoMetadata.GetValue<string>(Config);
+                        var eventArgs = new ChoiceChangedEventArgs(id, Array.IndexOf(options, value), value);
+                        invokeOnChangeEvents(modOptionMetadata, sender, eventArgs);
+                    }
+                    break;
+                case ChoiceAttribute choiceAttribute when memberInfoMetadata.ValueType == typeof(int):
+                    // index-based choice value
+                    {
+                        string[] options = choiceAttribute.Options;
+                        int index = memberInfoMetadata.GetValue<int>(Config);
+                        var eventArgs = new ChoiceChangedEventArgs(id, index, options[index]);
+                        invokeOnChangeEvents(modOptionMetadata, sender, eventArgs);
+                    }
+                    break;
+
+                case KeybindAttribute _:
+                    {
+                        var eventArgs = new KeybindChangedEventArgs(id, memberInfoMetadata.GetValue<KeyCode>(Config));
+                        invokeOnChangeEvents(modOptionMetadata, sender, eventArgs);
+                    }
+                    break;
+
+                case SliderAttribute _:
+                    {
+                        var eventArgs = new SliderChangedEventArgs(id, Convert.ToSingle(memberInfoMetadata.GetValue(Config)));
+                        invokeOnChangeEvents(modOptionMetadata, sender, eventArgs);
+                    }
+                    break;
+
+                case ToggleAttribute _:
+                    {
+                        var eventArgs = new ToggleChangedEventArgs(id, memberInfoMetadata.GetValue<bool>(Config));
+                        invokeOnChangeEvents(modOptionMetadata, sender, eventArgs);
+                    }
+                    break;
+            }
+        }
+
+        /// <summary>
+        /// Invokes the relevant method(s) specified with <see cref="OnChangeAttribute"/>(s)
+        /// and passes parameters when a value is changed.
+        /// </summary>
+        /// <typeparam name="TSource">The type of the event args.</typeparam>
+        /// <param name="modOptionMetadata">The metadata for the mod option.</param>
+        /// <param name="sender">The sender of the event.</param>
+        /// <param name="e">The event args from the OnChange event.</param>
+        private void invokeOnChangeEvents<TSource>(ModOptionAttributeMetadata<T> modOptionMetadata, object sender, TSource e)
+            where TSource : IModOptionEventArgs
+        {
+            if (modOptionMetadata.OnChangeMetadata == null)
+                return; // Skip attempting to invoke events if there are no OnChangeAttributes set for the member.
+
+            foreach (MemberInfoMetadata<T> onChangeMetadata in modOptionMetadata.OnChangeMetadata)
+            {
+                invokeEvent(onChangeMetadata, sender, e);
+            }
+        }
+
+        /// <summary>
+        /// Invoke the relevant method specified by a <see cref="ModOptionEventAttribute"/>
+        /// and passes relevant parameters.
+        /// </summary>
+        /// <typeparam name="TSource">The type of the event args.</typeparam>
+        /// <param name="memberInfoMetadata">The metadata for the method.</param>
+        /// <param name="sender">The sender of the event.</param>
+        /// <param name="e">The event args from the event.</param>
+        private void invokeEvent<TSource>(MemberInfoMetadata<T> memberInfoMetadata, object sender, TSource e)
+            where TSource : IModOptionEventArgs
+        {
+            if (!memberInfoMetadata.MethodValid)
+            {
+                // Method not found, log error and skip.
+                Logger.Error($"[OptionsMenuBuilder] Could not find the specified method: {typeof(T)}.{memberInfoMetadata.Name}");
+                return;
+            }
+
+            if (memberInfoMetadata.MethodParameterTypes is Type[] parameterTypes)
+            {
+                var parameters = new object[parameterTypes.Length];
+                var senderFound = false;
+                var eventArgsFound = false;
+                var modOptionEventFound = false;
+
+                for (var i = 0; i < parameterTypes.Length; i++)
+                {
+                    if (!senderFound && parameterTypes[i] == typeof(object))
+                    {
+                        senderFound = true;
+                        parameters[i] = sender;
+                    }
+                    else if (!eventArgsFound && parameterTypes[i] == typeof(TSource))
+                    {
+                        eventArgsFound = true;
+                        parameters[i] = e;
+                    }
+                    else if (!modOptionEventFound && parameterTypes[i] == typeof(IModOptionEventArgs))
+                    {
+                        modOptionEventFound = true;
+                        parameters[i] = e;
+                    }
+
+                    if (senderFound && eventArgsFound && modOptionEventFound)
+                        break;
+                }
+
+                memberInfoMetadata.InvokeMethod(Config, parameters);
+            }
+        }
+        #endregion
     }
 }
